extractor/requirements.txt
feedparser
requests
psycopg2-binary
python-dotenv
openai
apache-airflow-providers-postgres